# Implement unified CSS transform