--- a/tests/transforms/css/bundle.1.js
+++ b/tests/transforms/css/bundle.1.js
@@ -1,0 +1,4 @@
+/*eslint-env browser*/
+(() => {
+  document.querySelector('.js-blue');
+})();