--- a/tests/transforms/css/bundle.2.js
+++ b/tests/transforms/css/bundle.2.js
@@ -1,0 +1,4 @@
+/*eslint-env browser*/
+(() => {
+  document.getElementsByClassName('js-red');
+})();