--- a/tests/transforms/css/index.html
+++ b/tests/transforms/css/index.html
@@ -1,0 +1,8 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <style>/* CSS */</style>
+  </head>
+  <body class="red">
+  </body>
+</html>