--- a/tests/transforms/css/main.css
+++ b/tests/transforms/css/main.css
@@ -1,0 +1,17 @@
+/** CSS comment **/
+
+.red {
+  color: red;
+}
+
+.blue {
+  color: blue;
+}
+
+.js-blue {
+  color: blue;
+}
+
+.js-red {
+  color: red;
+}