--- a/tsconfig.json
+++ b/tsconfig.json
@@ -9,5 +9,6 @@
     "skipLibCheck": true,
     "lib": ["es2018", "DOM"],
   },
-  "include": ["**/*"]
+  "include": ["**/*"],
+  "exclude": ["tests/**"]
 }