--- a/types/eleventy/shortcodeContext.ts
+++ b/types/eleventy/shortcodeContext.ts
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2022 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+declare global {
+  export interface ShortcodeContext {
+    ctx: {
+      process: {
+        env: {[key: string]: string};
+      };
+      eleventy: {
+        version: string;
+        generator: string;
+        env: {[key: string]: string};
+      };
+      pkg: {
+        name: string;
+        version: string;
+        description: string;
+      };
+      locale: string;
+      title: string;
+      page: EleventyPage['page'];
+      collections: {[key: string]: EleventyCollectionObject};
+    };
+    page: EleventyPage['page'];
+  }
+}
+
+// empty export to keep file a module
+export {};