This PR tries to create a unified implementation for the CSS purify transforms living at

https://github.com/GoogleChrome/developer.chrome.com/blob/d8c4a0514b67863d2e38b738d38ee4e78c0c8b9c/site/_transforms/purify-css.js#L27-L34

and

https://github.com/GoogleChrome/web.dev/blob/725beb67e239556c4cbf600b2080ba4abad9dd7c/src/site/_transforms/purify-css.js#L37-L44

while implementing the best practices learnt in https://github.com/matthiasrohmer/developer.chrome.com/pull/2, like reading the CSS only once, minifying it once and also read the JS source ourselves instead of letting PurgeCSS do it as this would do a `fs.access` and `fs.readFile` call for every JavaScript file we have, for every page rendered (nuts!).

This is the API for usage:

```js
    eleventyConfig.addTransform('purifyCss', (new CssTransform()).configure({
      cssPath: 'dist/css/main.css',
      jsPaths: [
        'dist/js/**/*.js'
      ],
      insert: (content, result) => {
        return content.replace('</head>', `<style>${result}</style></head>`)
      }
    }));
```

Open tasks:

- [x] Compare build output on d.c.c
- [ ] Test with web.dev
- [x] Figure out if the staging environment differences on web.dev are still needed
- [x] Add tests (can CSS be read? is minified CSS smaller? can JS be read? does read JS have a length? is purged site smaller?)
- [x] Add safety nets (throw on empty CSS, throw on empty purged CSS)