@matthiasrohmer 2022-11-04
The decrease in build times is it worth though I think. All tests ran with csso 5.05 and purgecss 5.0.0, which are the recent versions and newer than what we currently have on `main`. The first run shows timings with all ELEVENTY_IGNORE flags set to true, the second timing is for a full production build.

**Old transformer:**
[11ty] Copied 721 files / Wrote 84 files in 6.67 seconds (79.4ms each, v1.0.1)
[11ty] Copied 721 files / Wrote 4038 files in 317.50 seconds (78.6ms each, v1.0.1)

Old transformer, full build, after [JS fix](https://github.com/GoogleChrome/developer.chrome.com/pull/4201)
[11ty] Copied 721 files / Wrote 4097 files in 406.03 seconds (99.1ms each, v1.0.1)

**New transformer:**
[11ty] Copied 721 files / Wrote 84 files in 5.20 seconds (61.9ms each, v1.0.1)
[11ty] Copied 721 files / Wrote 4038 files in 231.61 seconds (57.4ms each, v1.0.1)

New transformer, full build, after [JS fix](https://github.com/GoogleChrome/developer.chrome.com/pull/4201)
[11ty] Copied 721 files / Wrote 4097 files in 313.01 seconds (76.4ms each, v1.0.1)